refactor(foundation-demo): add tests for  daff-sidebar animations
