+++ apps/foundation-demo/src/app/design/molecules/sidebar/animation/sidebar-animation-state.ts
@@ -1,0 +1,11 @@
+export const getAnimationState = (open : boolean, enabled: boolean = true) => {
+    if(open && enabled){
+        return "open";
+    }
+    else if(open && !enabled) {
+        return "open-instant";
+    }
+    else {
+        return "void";
+    }
+}
